tests/submission/PhilipTaylor/canvas/2d.gradient.interpolate.zerosize.stroke.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.gradient.interpolate.zerosize.stroke</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.gradient.html">gradient</a>.<a href="index.2d.gradient.interpolate.html">interpolate</a>.<a href="index.2d.gradient.interpolate.zerosize.html">zerosize</a>.stroke</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.gradient.linear.zerosize">2d.gradient.linear.zerosize</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

var g = ctx.createLinearGradient(50, 25, 50, 25); // zero-length line (undefined direction)
g.addColorStop(0, '#f00');
g.addColorStop(1, '#f00');
ctx.strokeStyle = g;
ctx.rect(20, 20, 60, 10);
ctx.stroke();
_assertPixel(canvas, 19,19, 0,255,0,255, "19,19", "0,255,0,255");
_assertPixel(canvas, 20,19, 0,255,0,255, "20,19", "0,255,0,255");
_assertPixel(canvas, 21,19, 0,255,0,255, "21,19", "0,255,0,255");
_assertPixel(canvas, 19,20, 0,255,0,255, "19,20", "0,255,0,255");
_assertPixel(canvas, 20,20, 0,255,0,255, "20,20", "0,255,0,255");
_assertPixel(canvas, 21,20, 0,255,0,255, "21,20", "0,255,0,255");
_assertPixel(canvas, 19,21, 0,255,0,255, "19,21", "0,255,0,255");
_assertPixel(canvas, 20,21, 0,255,0,255, "20,21", "0,255,0,255");
_assertPixel(canvas, 21,21, 0,255,0,255, "21,21", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
